container: Stop pulling alternate stats when getting policy index
We include storage policy information in the response to container
server PUTs, including when stat updates from shards. This should
be a fairly cheap thing to include, but profiling revealed that
this could still take a half-second due to us using the general
get_info interface, which computes object count and bytes used
based on shard ranges. We weren't using this information anyway,
so let's just look at the container_stat table.

Change-Id: I43e4b01c41fe67c4556f5d9075d8cc8708c55182