slo: 500 if we can't load the manifest
When the proxy app is trying to send back an object and hits an error
(maybe a timeout, maybe an EC decode error) *after* it has sent headers
and started streaming data, it just stops sending data, expecting
clients to notice the discrepency in Content-Length and retry.

When that happened in SLO while reading a manifest, previously we'd just
assume the manifest is empty and send back an empty response. This would
cause confusion for users (who'd think we lost data or soemthing) and
was clearly wrong.

Now, return a 500 to the client. Retrying is perfectly reasonable.

Change-Id: I7fc923ad0ef37459b7a76ce360dd7f320053d3f7