[stable-only] Cap virtualenv/setuptools
With latest virtualenv release (20.24.0) and its bundled setuptools
version, stable/victoria gates started to fail [1], because setuptools
is not compatible with the pbr version (5.5.0) which is in victoria
branch's upper constraint [2]. pbr's next release (5.5.1) removed the
failing code, this is why the change is only needed in stable/victoria.

This patch caps virtualenv to <20.24.0 to fix stable/victoria's gate,
and also caps tox for the 'inner' virtualenv, otherwise tox>4 would be
installed, that would introduce failures (those failures are fixed in
newer stable branches, but it would need several backports and 'outer'
tox is capped <4 anyway, so no worth dealing with it here).

Also:
- Pin the pep8 and lower-constraints jobs to bionic nodes, which seem
  better positioned to deal with py27.
- Pin pyOpenSSL==21.0.0 in lower-constraints.txt, as that was the last
  version to support py27.

[1] AttributeError: module 'setuptools.command.easy_install' has no attribute 'get_script_header'
[2] https://opendev.org/openstack/requirements/src/commit/636b6b3bde4b68ef27ac6b524cb888152eddf517/upper-constraints.txt#L44

Change-Id: I05fdf0e1b00d54f5364d02c59019a4c54eb061ca