s3api: use string for swift info request method
A swift request was being constructed with its request method equal to
the s3api's ListingEtagMiddleware. This was benign because the swift
request is just used as a context to get container info, and its
method is never used. However, it is better to use a dummy string
value for the method. 'TEST' has been previously when an auth request
is needed prior to getting container info, so the same is employed
here.

Change-Id: Ie1c70a0ada9883cc4c16b7fc8c2d91a50f4a91b7