sharder: reduce rate of 'Reclaimable db...' warnings
Sharded DBs that have been completely emptied of objects will not be
removed if auto-sharding is disabled. The shards remain in place and
the sharder will repeatedly warn of a 'Reclaimable db stuck waiting
for shrinking'. This can occur often if expiring objects containers
become sharded, since these containers will always, by design, be
emptied. Each time this happens, another warning log starts to be
repeatedly emitted.

This patch reduces the rate of this specific warning to once per 24
hours.

Change-Id: I493cd661fe98817ca5fd2daf2a681fdf69e28923