proxy: emit metric when updating namespaces memcache set fails
Previously, if a memcache set command had an unexpected response then
an error was logged but no exception raised from MemcacheRing.set.
With this patch, when MemcacheRing.set fails in this way, a
MemcacheConnectionError will be raised if 'raise_on_error' is
True. This exception is now handled when the proxy tries to set
updating namespaces in memcache, and a counter metric emitted with
name 'object.shard_updating.cache.set_error'. If the set is successful
an 'object.shard_updating.cache.set' counter metric is emitted.

The logging of MemcacheRing errors is improved to include the memcache
key being set, in line with exception logging.

Change-Id: Icfc97751c80f4bb4a3373796c01bff6ed5843937