proxy: remove x-backend-record-type=shard in object listing
When constructing an object listing from container shards, the proxy
would previously return the X-Backend-Record-Type header with the
value 'shard' that is returned with the initial GET response from the
root container. It didn't break anything but was plainly wrong.

This patch removes the header from object listing responses to request
that did not have the header. The header value is not set to 'object'
because in a request that value specifically means 'do not recurse
into shards'.

Change-Id: I94c68e5d5625bc8b3d9cd9baa17a33bb35a7f82f