Update malformed_example.db to actually be malformed
Seems since somewhere around sqlite 3.40+ our in tests malformed sqlite
db isn't malformed anymore. I don't actually know how it was malformed
but looking in a hex editor it seems to have a bunch of null
truncated in the middle of the file. Which maybe isn't an issue anymore?

Instead I've gone and messed up what looks like to be the marker before
defining the test table data at the end of file, so from:

  00001FF0  00 00 00 00 00 00 00 00 00 00 00 03 01 02 0F 31 ...............1
                                             ^^

To:

  00001FF0  00 00 00 00 00 00 00 00 00 00 00 FF 01 02 0F 31 ...............1
                                             ^^

Basically FF'ed the start of the data marker (at least what I'm calling
it).

Closes-Bug: #2051067
Change-Id: I2a10adffa39abbf7e97718b7228de298209140f8