proxy: Add list of backend response data to request environ
Middleware may want to inspect the attributes of backend responses,
for example to find interesting headers that backends have
returned. When there is a multiplicity of backend responses, the
headers from all of them cannot be added to the single proxy
response. Besides, it is not always appropriate for backend headers to
be added to the proxy response. This patch therefore adds a list of
backend response attributes to the request environ under the key
'swift.backend_responses'.

This patch refactors the proxy controllers to replace the multitude of
lists holding statuses, reasons, bodies, headers with a single list of
ResponseData that encapsulates the attributes of each backend
response. This single list to Controller.best_response(), where it is
added to the request environ.

Drive-by: Stop passing server_type to best_response because it is
already an attribute of Controller. Always include the request method
when logging an error if quorum is not met for a backend response
status.

Change-Id: Ibd70d9b631075a23bff258c4f6f55ca4df59c0c2