Clean up centos-7 based jobs
There is no Centos-7 nodeset anymore, so let's get rid of those jobs,
or use centos-8-stream if possible. Also remove old rolling upgrade
jobs that would run against branches that are already End of Life.
Grenade job is currently broken, let's remove it from gate and set it
non-voting on check queue to unblock gate. Grenade is part of the
integrated-gate-object-storage template [1], so that has to be removed
and its jobs have to be added explicitly (definition of job
tempest-integrated-object-storage is already re-defined (overriden),
but openstacksdk-functional-devstack has to be added).

[1] https://opendev.org/openstack/tempest/src/commit/1c3fc43c9dae274a19683b003f78b5f50edbc062/zuul.d/integrated-gate.yaml

Change-Id: Id56057262fc37add3a47c50946067f8c873c7515