slo: use pre-auth'd request to fetch segments from reserved namespace
There's precedent in symlink.py for using a pre-auth'd request to
resolve a symlink when it points to a target in the reserved
namespace. This patch changes the SLO segment iterator to do similar
i.e. if the SLO manifest has a header indicating that reserved names
are allowed, then it is assumed that another middleware created the
manifest, determined that the segments are to be in the reserved
namespace, and takes responsibility for managing access to them.  In
this case the SegmentedIterable will use pre-auth'd requests to fetch
the segments.

Change-Id: I8689da307dec96eec9ff93f6b15938e6374e3b79