obj: Have handoff_delete reduce replication to handoffs
Suppose you've got a full cluster. Some time after bringing in more
hardware and setting handoff_delete=1 and handoffs_first=true but
before you're all the way out of the woods, you notice one of the
partitions has one primary on fresh hardware and at least one primary
with a full disk.

The fresh-hardware primary should have few if any handoffs; as a
result, it goes back to syncing mode. For every 507 it gets back, it
finds a handoff to sync to. Those handoffs are going to try to clean
that up as quick as they can, and with handoff_delete=1, they only
need to sync with the fresh-hardware primary in order to delete --
all we've done is waste some rsync connections and network bandwidth,
and the fresh-hardware primary will try to do it again.

Now, only be willing to sync to at most handoff_delete-1 handoffs.
This respects the operator's stated willingness to tolerate reduced
durability for the sake of faster replication and ensures that no
handoff should receive a bunch of data just to immediately delete
it the next chance it gets.

Note that the intentionally-create-handoffs behavior can now be
disabled entirely via handoff_delete = 1.

Change-Id: I4bbd28a4c8554cba742a68028ad1aa24d897146f