Stop using cgi.parse_header
With py311 we started seeing warnings like

   DeprecationWarning: 'cgi' is deprecated and slated for removal
   in Python 3.13

The recommended replacement isn't quite up to snuff (doesn't handle
multiple parameters), but we already have a reasonably close replacement.

Add a "loose" mode to parse_content_type to make it suitable when there
may be a slash in a parameter token. Add a new utils.parse_header function
that takes advantage of the parse_content_type "loose" mode.

Closes-Bug: #2084472
Change-Id: Ie281ff90796f2d68840952c95669f264480b1b4c