mpu: complete mpu with user content-type when available
If the user sent a content-type header with a createUpload request
then, during completeUpload, MPU middleware sends that with the
manifest PUT and the symlink PUT.

If the user did not send a content-type header with a createUpload
request then, during completeUpload, MPU middleware sends a default
'application/x-mpu' with the manifest PUT and sends no content-type
with the symlink PUT. The proxy app will guess content-type for the
symlink based on the user namespace path.

Also clarify some module constant names.

Change-Id: Ib3083a7e987c0e6ed921c68b44f074aebd4af5c3