versioning: 411 PUTs with neither content-length nor transfer-encoding
... just like we would do in a normal container. Previously, we'd try to
read a byte from the client which, due to a bug in eventlet HTTP framing,
would either hang until we hit a timeout or worse read from the next
pipelined request.

This required that we reset a (repeatedly-reused!) request in s3api to
have an empty body, or it would start triggering 411s, too.

See also: https://github.com/eventlet/eventlet/pull/985

Closes-Bug: #2081103
Change-Id: I56c1ecc4edb953c0bade8744e4bed584099f29c7