mpu: check for manifest in versions container during parts audit
The mpu middleware is careful to NOT write down delete-markers while
versioning is enabled. However, there's a gap when:

  * versioning is enabled
  * an mpu manifest is overwritten
  * the mpu manifest is copied to the versions container
  * not all copies of the mpu manifest are deleted in the user namespace
    and a stale replica is left in the user namesapce
  * versioning becomes disabled
  * the current object in the user namespace is overwritten
  * the mpu middleware sees that the stale replica was deleted
  * the mpu middleware writes a delete-marker - BAD!

To mitigate for this the mpu-auditor will now make a HEAD request to
check if the manifest is in the versions container before deleting the
parts.

To enable this HEAD request to be made it must be possible to
determine the potential version object name from the upload name
i.e. for the version-id to be determined from the upload-id. This is
possible if the version-id is always derived from the created-at time
of the mpu (i.e. the data-timestamp) which is embedded in the
upload-id. The object_versioning middleware is therefore modified to
use the data-timestamp rather than the x-timestamp when creating a
version-id for an object copied from the user namespace to the
versions container.

Change-Id: I63e9027ebe855774da90b2c435f3393a6db2bc79